requirements.txt
python-whois
selenium
requests